신규거래 크롤링 방식 수정
