+++ apps/jibsayo/src/app/api/shared/libs/fireStore.ts
@@ -1,0 +1,21 @@
+import { AdminFirestoreClient } from '@polymorph/firebase';
+
+import { COLLECTIONS } from '../../consts';
+
+export const firestoreClient = new AdminFirestoreClient({
+  collectionName: COLLECTIONS.NEW_TRANSACTIONS_CACHE,
+  projectId: process.env.FIREBASE_PROJECT_ID,
+  // 서비스 계정 정보 (환경변수에서 읽기)
+  serviceAccount: {
+    type: 'service_account',
+    project_id: process.env.FIREBASE_PROJECT_ID,
+    private_key_id: process.env.FIREBASE_PRIVATE_KEY_ID,
+    private_key: process.env.FIREBASE_PRIVATE_KEY?.replace(/\\n/g, '\n'),
+    client_email: process.env.FIREBASE_CLIENT_EMAIL,
+    client_id: process.env.FIREBASE_CLIENT_ID,
+    auth_uri: 'https://accounts.google.com/o/oauth2/auth',
+    token_uri: 'https://oauth2.googleapis.com/token',
+    auth_provider_x509_cert_url: 'https://www.googleapis.com/oauth2/v1/certs',
+    client_x509_cert_url: process.env.FIREBASE_CLIENT_CERT_URL,
+  } as any,
+});
