+++ apps/jibsayo/src/app/api/shared/utils/text.ts
@@ -1,0 +1,6 @@
+export const obfuscateKorean = (text: string): string => {
+  return Buffer.from(text)
+    .toString('base64')
+    .replace(/\s/g, '_')
+    .replace(/[+/=]/g, '');
+};
